Address MATLAB RT converter review feedback
